Добавлены Loop, Loopq, Clear, Remove(id-id).

+++ DiscordMusicBot/src/main/java/com/djusufcompany/discordmusicbot/TrackScheduler.java
@@ -1,7 +1,6 @@
 package com.djusufcompany.discordmusicbot;
 
 
-import com.github.kiulian.downloader.Config;
 import com.github.kiulian.downloader.YoutubeDownloader;
 import com.github.kiulian.downloader.downloader.request.RequestVideoFileDownload;
 import com.github.kiulian.downloader.downloader.request.RequestVideoInfo;
@@ -28,7 +27,9 @@ public class TrackScheduler extends AudioEventAdapter
     private ArrayList<TrackInfo> queue;
     private ArrayList<Response<File>> downloaded;
     private Integer currentTrackNumber = 0;
-    private Integer downloadBuferNumber = 3;
+    private Integer downloadReserve = Integer.valueOf(ConfProperties.getProperty("downloadReserve"));
+    private Boolean isTrackLooped = false;
+    private Boolean isQueueLooped = false;
 
     public TrackScheduler(AudioPlayer player, Guild guild)
     {
@@ -38,24 +39,22 @@ public class TrackScheduler extends AudioEventAdapter
         this.downloaded = new ArrayList<>();
     }
 
-    public void addToQueue(String tracksUrl)
+    public void addToQueue(String url)
     {
         YoutubeDownloader downloader = new YoutubeDownloader();
-        String videoId = urlToId(tracksUrl);
-        RequestVideoInfo requestInfo = new RequestVideoInfo(videoId);
-        Response<VideoInfo> responseInfo = downloader.getVideoInfo(requestInfo);
-        VideoInfo video = responseInfo.data();
-        
-        queue.add(new TrackInfo(video.details().author(), video.details().title(), tracksUrl));
+        VideoInfo video = getVideoInfo(downloader, url);
 
-        if (downloaded.size() < downloadBuferNumber)
+        queue.add(new TrackInfo(video.details().author(), video.details().title(), url));
+
+        if (downloaded.size() < downloadReserve)
         {
-            downloaded.add(loadTrackFromUrl(tracksUrl));
+            downloaded.add(loadTrackFromUrl(url));
         }
         if (downloaded.size() == 1)
         {
-            PlayerManager.getInstance().loadAndPlay(guild, downloaded.get(0).data().getAbsolutePath());
+            nextTrack();
         }
+        System.out.println("Add CurrentTrackNumber: " + currentTrackNumber);
     }
 
     public void play(AudioTrack track)
@@ -66,34 +65,72 @@ public class TrackScheduler extends AudioEventAdapter
     @Override
     public void onTrackEnd(AudioPlayer player, AudioTrack track, AudioTrackEndReason endReason)
     {
-        if (endReason.mayStartNext)
+        if (isTrackLooped)
         {
+            playTopTrack();
+        }
+        else
+        {
+            currentTrackNumber += 1;
+            downloaded.get(0).data().delete();
+            downloaded.remove(0);
             nextTrack();
         }
-        currentTrackNumber += 1;
     }
 
     public void nextTrack()
     {
-        if ((downloaded.size() >= downloadBuferNumber) && (queue.size() - currentTrackNumber > downloadBuferNumber))
+        if (currentTrackNumber + downloaded.size() < queue.size())
         {
-            downloaded.add(loadTrackFromUrl(queue.get(currentTrackNumber + downloadBuferNumber).getUrl()));
+            //System.out.println("Загрузка трека из !internet!");
+            //System.out.println("Размер очереди: " + queue.size() + "; Индекс: " + (currentTrackNumber + downloadReserve));
+            //System.out.println("URL: " + queue.get(currentTrackNumber + downloadReserve - 1).getUrl());
+            Response<File> loadTrackFromUrl = loadTrackFromUrl(queue.get(currentTrackNumber + downloadReserve - 1).getUrl());
+            //System.out.println("Добавление трека в загруженные");
+            downloaded.add(loadTrackFromUrl);
         }
-        downloaded.get(0).data().delete();
-        downloaded.remove(0);
+
         if (downloaded.size() == 0)
         {
-            guild.getAudioManager().closeAudioConnection();
+            //System.out.println("Размер буфера равен нулю");
+            if ((isQueueLooped) && (queue.size() > 0))
+            {
+                //System.out.println("Очередь зациклена");
+                ArrayList<TrackInfo> tmpQueue = (ArrayList<TrackInfo>) queue.clone();
+                //System.out.println("Очередь скопирована");
+                Integer size = PlayerManager.getInstance().getMusicManager(guild).scheduler.getQueueInfo().size();
+                //System.out.println("Получение размера очереди: " + size);
+                for (int i = 0; i < size; i += 1)
+                {
+                    PlayerManager.getInstance().getMusicManager(guild).scheduler.remove(0);
+                    //System.out.println("Удаление " + i + "-го элемента. Количество элементов очереди: " + queue.size());
+                }
+                for (int i = 0; i < tmpQueue.size(); i += 1)
+                {
+                    addToQueue(tmpQueue.get(i).getUrl());
+                    //System.out.println("Добавление " + i + "-го элемента. Количество элементов очереди: " + queue.size());
+                }
+                //System.out.println("nextTrack: currentTrackNumber: " + currentTrackNumber);
+            }
+            else
+            {
+                guild.getAudioManager().closeAudioConnection();
+            }
         }
         else
         {
-            PlayerManager.getInstance().loadAndPlay(guild, downloaded.get(0).data().getAbsolutePath());
+            playTopTrack();
         }
     }
 
     public TrackInfo getTrackInfo()
     {
         return queue.get(currentTrackNumber);
+    }
+
+    public void playTopTrack()
+    {
+        PlayerManager.getInstance().loadAndPlay(guild, downloaded.get(0).data().getAbsolutePath());
     }
 
     public ArrayList<TrackInfo> getQueueInfo()
@@ -103,68 +140,70 @@ public class TrackScheduler extends AudioEventAdapter
 
     public void skip()
     {
-        if (currentTrackNumber + downloadBuferNumber - 1 < queue.size())
-        {
-            downloaded.add(loadTrackFromUrl(queue.get(currentTrackNumber + downloadBuferNumber).getUrl()));
-        }
-        downloaded.get(0).data().delete();
-        downloaded.remove(0);
         player.stopTrack();
-        if (downloaded.size() == 0)
-        {
-            guild.getAudioManager().closeAudioConnection();
-        }
-        else
-        {
-            PlayerManager.getInstance().loadAndPlay(guild, downloaded.get(0).data().getAbsolutePath());
-        }
+        nextTrack();
     }
 
     public void remove(Integer id)
     {
-        if ((id >= queue.size()) || (id < 0))
+        System.out.println("Удаление трека с индексом " + id);
+        if ((id >= 0) && (id < queue.size()))
         {
-            System.out.println("1");
-            return;
-        }
-        queue.remove(queue.get(id));
-        if (id == currentTrackNumber)
-        {
-            System.out.println("2");
-            currentTrackNumber -= 1;
-            skip();
-        }
-        else if (id < currentTrackNumber)
-        {
-            System.out.println("3");
-            currentTrackNumber -= 1;
-        }
-        else if (currentTrackNumber + downloadBuferNumber > id)
-        {
-            System.out.println("4");
-            downloaded.get(id - currentTrackNumber).data().delete();
-            downloaded.remove(id - currentTrackNumber);
-            if (currentTrackNumber + downloadBuferNumber - 1 < queue.size())
+            System.out.println("id корректный");
+            queue.remove(queue.get(id));
+            System.out.println("Элемент удалён из очереди");
+            if (id == currentTrackNumber)
             {
-                downloaded.add(loadTrackFromUrl(queue.get(currentTrackNumber + downloadBuferNumber).getUrl()));
+                System.out.println("Тип удаления 1");
+                //currentTrackNumber += 1;
+                downloaded.get(0).data().delete();
+                System.out.println("Удаление из очереди элемента " + (id - currentTrackNumber));
+                downloaded.remove(0);
+                System.out.println("Переход к следующему треку");
+                skip();
+                
+//                downloaded.get(id - currentTrackNumber).data().delete();
+//                downloaded.remove(downloaded.get(id - currentTrackNumber));
+//                if (downloaded.size() != 0)
+//                {
+//                    currentTrackNumber -= 1;
+//                }
+//                skip();
+//                System.out.println("remove 1 тип currentTrackNumber: " + currentTrackNumber);
+            }
+            else if (id < currentTrackNumber)
+            {
+                System.out.println("Тип удаления 2");
+                currentTrackNumber -= 1;
+                System.out.println("remove 2 тип currentTrackNumber: " + currentTrackNumber);
+            }
+            else if (currentTrackNumber + downloadReserve > id)
+            {
+                System.out.println("Тип удаления 3");
+                downloaded.get(id - currentTrackNumber).data().delete();
+                downloaded.remove(downloaded.get(id - currentTrackNumber));
+                if (currentTrackNumber + downloadReserve - 1 < queue.size())
+                {
+                    downloaded.add(loadTrackFromUrl(queue.get(currentTrackNumber + downloadReserve - 1).getUrl()));
+                }
+
+                
+
+                System.out.println("remove 3 тип currentTrackNumber: " + currentTrackNumber);
             }
         }
     }
 
     private Response<File> loadTrackFromUrl(String url)
     {
+        System.out.println("Загрузка видео по адресу: " + url);
         YoutubeDownloader downloader = new YoutubeDownloader();
-        Config config = downloader.getConfig();
-        config.setMaxRetries(0);
-        String videoId = urlToId(url);
-
-        RequestVideoInfo requestInfo = new RequestVideoInfo(videoId);
-        Response<VideoInfo> responseInfo = downloader.getVideoInfo(requestInfo);
-        VideoInfo video = responseInfo.data();
+        VideoInfo video = getVideoInfo(downloader, url);
 
         Format format = video.bestAudioFormat();
         if (format == null)
         {
+            System.out.println("У трека нет аудио-версии");
             format = video.bestVideoFormat();
         }
         if (format != null)
@@ -173,14 +212,18 @@ public class TrackScheduler extends AudioEventAdapter
             try
             {
                 outputDir = Files.createTempDirectory("video").toFile();
+
             }
             catch (IOException ex)
             {
-                Logger.getLogger(TrackScheduler.class.getName()).log(Level.SEVERE, null, ex);
+                Logger.getLogger(TrackScheduler.class
+                        .getName()).log(Level.SEVERE, null, ex);
             }
+            System.out.println("Папка для вывода: " + outputDir.getAbsolutePath());
             RequestVideoFileDownload requestFile = new RequestVideoFileDownload(format)
                     .saveTo(outputDir)
                     .renameTo("video");
+            System.out.println("Запрос");
             return downloader.downloadVideoFile(requestFile);
         }
         return null;
@@ -189,28 +232,54 @@ public class TrackScheduler extends AudioEventAdapter
     private String urlToId(String url)
     {
         String videoId = null;
-        if (url.contains("&t="))
+
+        String[] tagArray =
         {
-            url = url.substring(0, url.indexOf("&t="));
-        }
-        else if (url.contains("?t="))
+            "&t=",
+            "&t=",
+            "&list="
+        };
+        String[] urlTypeArray =
         {
-            url = url.substring(0, url.indexOf("?t="));
-        }
-        if (url.contains("&list="))
+            "https://www.youtube.com/watch?v=",
+            "https://youtu.be/"
+        };
+
+        for (String tag : tagArray)
         {
-            url = url.substring(0, url.indexOf("&list="));
+            if (url.contains(tag))
+            {
+                url = url.substring(0, url.indexOf(tag));
+            }
         }
 
-        if (url.contains("https://www.youtube.com/watch?v="))
+        for (String urlType : urlTypeArray)
         {
-            videoId = url.substring("https://www.youtube.com/watch?v=".length());
+            if (url.contains(urlType))
+            {
+                videoId = url.substring(urlType.length());
+            }
         }
-        else if (url.contains("https://youtu.be/"))
-        {
-            videoId = url.substring("https://youtu.be/".length());
-        }
+
         return videoId;
+    }
+
+    public void changeTrackLoopMode()
+    {
+        isTrackLooped = !isTrackLooped;
+    }
+
+    public void changeQueueLoopMode()
+    {
+        isQueueLooped = !isQueueLooped;
+    }
+
+    public VideoInfo getVideoInfo(YoutubeDownloader downloader, String url)
+    {
+        String videoId = urlToId(url);
+        RequestVideoInfo requestInfo = new RequestVideoInfo(videoId);
+        Response<VideoInfo> responseInfo = downloader.getVideoInfo(requestInfo);
+        return responseInfo.data();
     }
 }
